HYD-5480 Chroma agent causes constant UDEV change events
On systems running the chroma-agent process, "udevadm monitor" reports constant
UDEV change events for all disk drives (/dev/sd*) and device mapper drives
(/dev/dm-*). The events come in bursts about every one second.

Stopping the chroma-agent process with "/etc/init.d/chroma-agent stop" causes
the events to stop.

These constant change events are likely not good for the storage subsystem and
could affect Lustre performance negatively.

This was reported by HP as HP-44

The solution is to code more defensively when dealing with multimap devices. It
is quite possible that a multimap devices reports a 'child' device that is not
actually present in the system. Whilst debugging this it was also discovered that
the agent is not resilient to lvm2 not being installed and so the code was updated
to make that more resilient.

 - Check that devices are reported by blkid before trying to add them as children of
   the parent multimap device.
 - Deal gracefully with the case of vgs or lvg not being installed.

Change-Id: I277504cdcaaf6c01b84ea1bbeea2c20159207275
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18423
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>