HYD-5596 Removing last server in corosync cluster fails on el7
When performing action "Unconfigure corosync" on a node which happens
to be the only member of a corosync cluster, "Error: Unable to update
any nodes" is encountered.

This occurs because pcs is attempting to contact other nodes in the
cluster to notify them of the local node removal.

Solution is to check the corosync configuration to see if the local
node is the only member of the cluster, and if it is then return
successful rather than returning this error.

  - update manage_corosync2 to not return the aforementioned error
    if the node is the only member of the cluster
  - remove code in the same file that explicitly removes corosync.conf
    as this is done implicitly by 'pcs remove node' command
  - update test_corosync_config.py
  - amend comments in chroma_agent/lib/shell.py

Change-Id: I499d39f11c58acffecb91cdf7418aa09a9b5f6a9
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/18822
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>