HYD-5675 File system creation failed, because primary server not locatable
The MGT is started on the secondary server, because the primary server is
not locatable. The symptom of this issue is that it doesn't register the 
MDT/OST because the MGS is started on the secondary. This is a Lustre
limitation and believe with Primary/Secondary model, not a bug but a fact 
that you have to be on the primary MGS to add MDT/OSTs.

The root cause is that between  el6 and el7 (truthfully we should probably
be looking at Pacemaker or Corosync versions but as of today we don't know
which version the switch occurs at) Pacemaker started to use fully qualified
domain names rather than just the nodename.  
lotus-33vm15.lotus.hpdd.lab.intel.com vs lotus-33vm15. To keep compatibility 
easily we have to make the constraints follow the same fqdn vs node.

 - Use fqdn names for el7 and nodename of el6.
 - Update an error in node_admin.py I found when copying the code
   from nodeadmin. The existing code only works (its an import)
   because of where the file happens to sit in the hierarchy

Change-Id: I2f0a25c1f84e14a4ab4989629f156493653556d8
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19063
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>