HYD-5694 VGS missing causes a traceback to occur
The lvm code does not correctly detect when vgs is not installed
resulting in a traceback in the code and potential failure to
discover some devices.

The root cause is a misimplementation of HYD-5480 which in return
relied on a misimplementation of unit tests that in correctly
emulated a file-not found error for a missing command. The solution
is to correct the unit test framework to correctly raise an OSError
when a command is not found (command_capture_testcase.py) and then
fix all the unit test errors that occur.

The init test fixes are entirely test fixes other than the vgs/lvs
case where a code change is required.

 - Update command_capture_testcase.py to correctly handle command
   not found.
 - Update tests for zfs, corosync to correctly test for file not
   found
 - Update device_mapper and test to deal with vgs/lvs not being
   found

Change-Id: Ica4f3d87766c98f090da078f7a8b02dcb5784c26
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19085
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>