HYD-5772 Enable mcast port for corosync on EL7
Currently it is not possible to change the multicast port in the
corosync configuration for an EL7 managed server using the IML UI.

Because of the differences in the management of corosync and
pacemaker on EL7 distro's, namely that we use pcs to manage cluster
configuration and setup, we require a different approach to changing
the multicast port in corosync.

The corosync.conf file on the managed server selected in the UI
is updated with the new mcast port (for both ring 0 & 1 interfaces).
Corosync polls the config file and reads the updated value.

If mcast value is different from the value on the ha pair, corosync
will raise an alert until the values are made the same.

  - update original configure_corosync to remove firewall rule
    for old mcast port when changing to new value
  - rename configure_corosync2 to change_mcast_port in agent code
    which receives old and new ports, edits corosync.conf with
    new mcast port value, remove firewall rule for old port and
    add rule for the new port, run step from ConfigureCorosync2Job
  - modify AutoConfigureCorosyncStep and test_manual_ring1_config to
    pass configure_corosync() 4 params (including old and new mcast
    port values)
  - update simulator fake_server.py with amended configure_corosync
    and change_mcast_port methods

Change-Id: I4359881bac67d73b97b90cbe2ca247574a8faa51
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/19445
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>