HYD-6077 Remove python-dateutil from IML code
Python-datautil is blacklisted AND it turns out the latest version doesn't
support this piece of code

from dateutil.tz import tzlocal

self._old_std_offset = tzlocal._std_offset
_std_offset is not a member of tzlocal

By removing dateutil we can kill the proverbial two birds with one stone.

The method is to create an IMLDateTime class which is a descendant of pythons
datetime. It adds some useful functionality such as a fairly general date
parser, a method for getting the local time and utc time with the timezone
present. Unit tests are added.

This patch does not remove dateutil from chroma-externals because it is
required by tastypie. This dependency will be resolved separately.

 - Add IMLDateTime with parser, utcnow and now functionality.
 - Add to timezone classes FixedOffset which gives a timezone at a fixed offet
   of minutes and LocalOffset which gives a timezone at the offset of the local
   timezone.

Change-Id: I095836791c874a63f8527ad5afbc0908e92c3cb8
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/20811
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>