HYD-5135 failed to fail back target testfs-OST
The failover / failback for resources will sometimes fail to occur.
Althought the command is issue correctly pacemaker does not actually
action the device. This occurs when an error is present on the target,
and can be seen from the crm_mon -1 command.

crm_mon -1 -r
Last updated: Thu Nov 19 04:16:27 2015
Last change: Thu Nov 19 04:16:27 2015
Stack: classic openais (with plugin)
Current DC: lotus-8vm8 - partition with quorum
Version: 1.1.11-97629de
2 Nodes configured, 2 expected votes
3 Resources configured

Online: [ lotus-8vm7 lotus-8vm8 ]

Full list of resources:

 st-fencing	(stonith:fence_chroma):	Started lotus-8vm8
 testfs-OST0000_cc953e	(ocf::chroma:Target):	Started lotus-8vm8
 testfs-OST0001_b7d34d	(ocf::chroma:Target):	Started lotus-8vm8

Failed actions:
    testfs-OST0000_cc953e_monitor_5000 on lotus-8vm7 'not running' (7): call=19, status=complete, last-rc-change='Thu Nov 19 04:16:03 2015', queued=0ms, exec=263ms

Manually logging into a node a clearing the error with (in this case)

crm_resource --resource testfs-OST0000_cc953e --cleanup

will then either cause the command to complete or the next command issued
to execute correctly.

The problem seems to be that when a node is rebooted pacemaker will consider it
'node running' for longer than might be expected and so the cleanup forces pacemaker
to sync-up with the nodes.

The implemented solution is to cleanup before moving resources but also execute a
crm_mon -1 before doing the cleanup so that in the event of issues the state before
the actions can be seen.

 - Add crm_mon -1 before any actions take place in the resource move
 - Add crm_resource cleanup before moving the devices
 - Add unit test code to cover the pass and fail cases
 - Some minor related code cleanup that was discover (no functional changes)

Change-Id: Ibc78bb5c1576387ff37b977bd7509d80d26d6ca4
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/20585
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>