HYD-6143 ConfigureCorosyncStep fails on EL7.2 when NetworkManager is down
Although the agent deals with the Network Manager being uninstalled it does
not deal with it being disabled. When disable the nmcli commands returns
an exit code of 8 when uninstalled an exit (sic) code of 127 is received.

Add detection of exit codes 127 and 8 in the unmanage_network command on the
agent.

 - Detect 127 and 8 as acceptable outcomes for the nmcli command
 - Add simple tests for success, disabled and uninstalled.

Change-Id: I75fb669e3753f39b5ab0591ed423eb2d68b89295
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/21035
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>