HYD-6261 Allow enumeration of plugin resource by partial attribute
It would be useful to be able to search for a plugin resource by a partial
attribute such as host_id or uuid. This is a requirement for HP and a
requirement for some of the zfs work.
A construct that allows

 for resource in self.find_by_attr(CLASS, attrib=value):
   print 'resource is ' % resource

where the attrib might be only one of and so not a unique identifier for the
class of resource would be useful.

The implementation of this is that attrib must be one of the identifier (key)
attributes of the resource and then the iterator can select any subset.

As a working example this change allows the lnet updates to move away from the
overly complicated sending of deletes in the data from the agent to a more
regular remove those that the agent didn't report any more strategy.

Because those changes are very well tested by inference the functionality of
find_by_attr is tested.

Comments and relayout of the BaseStoragePlugin code so that it is easier to
understand what functions are private, internal and public. This comments can
then be shared with customers including HP.

 - Change linux_network to return just the current nids and interfaces rather
   than the current and the no-deleted list
 - Add find_by_attr to storage_plugin
 - Document functions withing the BaseStoragePlugin so that the documents can
   be shared with HP and other customers
 - Layout BaseStoragePlugin to group functions and make private methods
   private using the powerful leading underscore technique
 - Update storage_plugin to allow searching of indentifiers where one of the
   values is missing. I.e. a partial match algoritim
 - Change linux-network manager plugin to delete nids/interfaces no longer sent
   by the agent.
 - Update tests to not send active/inctive list of nids/interfaces, but to just
   send the current list

Change-Id: I92469bf3c4404382058f5b6e6a6e236047609e49
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/21696
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>