HYD-6325 Purge filesystem uses debugfs which does not work with ZFS
When we remove a filesystem we purge its memory from the mgs. This is currently
done using debugfs however debugfs does not work with zfs and so if the mgs
(not the target) is a zfs target then the purge will fail.

The effect of this is that you cannot create a filesystem on an mgs that is the
same name as a previously deleted filesystem.

Talking with Andreas the use of debugs might be unnecessary (and a strange
approach). He suggested why would you use debugfs instead of just mounting
the MGT device directly as type "ldiskfs" and use "rm"?

This patch splits the functionality between zfs and ldiskfs with the ldiskfs
code staying unchanged (for now?) and the zfs code using the new method
proposed by Andreas.  A further patch or a repush of this patch should make
ldiskfs use the new method.

 - Add purge_filesytem_configuration to BlockDevice class
 - Pass type information with purge_filesystem command
 - Refactor the code to correctly unmount/remount the mgs
 - Add code to make sure the mgs device is available to the node which will
   purge the filesystem information.

Change-Id: I5b8d1a5cd01b554dedbee0e477bc3611675dbaf3
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/21902
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>