HYD-6478 HYD-5826 has a bug that means devices present are marked as not present
DeviceHelper._dev_major_minor has a bug because a device that does not have
S_ISBLK set in its st_mode is added to the non_existent_path list. Next time it
is accessed it is removed from the non_existent_paths list before being added,
removed ....

This is a bug but can also lead to confusion for someone trying to debug
something else. In the patch I added some debug for when a path is added or
removed and this debug highlighted the bug.

The fix is to treat the not responding path differently from the reponded but
not S_ISBLK in st_mode present path.

 - Add debug to DeviceHelper._dev_major_minor to show when paths are added or
   removed.
 - Fix up exit path so that paths are only added when they do not respond.

Change-Id: I4abae6a565fa3153ef5053cf27a7d462727f7056
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22485
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>