HYD-6489 Update method in HYD-6340 is overly simplistic and will cause field issues.
The update method used by HYD-6340 to ensure a change on one node is reflected
on another is overly simplistic and synchronous meaning that we can see long
outstanding commands stack up when changes occur.

We need a model where a flag is set on the agent which causes the next poll to
always update the manager. The Linux plugin can then use this to trigger
updates on other nodes.

The agent command will be force_plugin_update [plugin] name.

 - Update agent plugsin to respond to force_plugin_update, many of the plugins
   use the delta function anyway and so detection in the delta is sufficient.
 - Update agent linux plugin to respond to force_plugin_update.
 - Change the manager linux plugin to use the new trigger_plugin_update call
   and by inference the TriggerPluginUpdate(Job/Step).
 - Add the beginning of a new class to allow testing of jobs and steps. This
   brings in some existing code whilst adding the first Job tests in the
   product. The invoke capture code was existing.

Change-Id: I0b3ac07e20d7cb58dcc8b1259a801c553416fe29
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22557
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>