HYD-6471 IML fails to install on RH7.2 if network manager is not installed
The IML deals gracefully with network manager not being install by
looking for a return code of 127 (not found). However it turns out
that process returns an error code of 2 when an application is not
found.Because of this the configure corosync set fails.

As we have a range of error codes to fix this add the return code of
2 to the expected return codes.

    -Add return code 2 to list within unmanaged_network function
    -Edit test for node admin to include error code of 2

Change-Id: I2deea8512893a78df423724b7bf538f470f8d2f8
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/22492
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>