HYD-6525 ZFS Import will require -f during failover operations.
Ticket HYD-5037 implemented the hostid feature of zfs which provides 
some protection against a user or application accidently importing a 
zpool simultaneously on different hosts.

The net effect of this is that if a node that crashes or is fenced
whilst it has a zfs pool imported then zfs will prevent another node
from importing that pool. To overcome this we must uses the -f option
to import of the ZFS pool whilst ignoring the host information that may
exist in the pool.

The -f option must only be used for pacemaker operations as other 
operations should fail. For example if a bug in IML causes IML (as
opposed to pacemaker) to attempt to import a zpool that is import
elsewhere then that operation should fail.

This is justifiable because we trust pacemaker to not try and mount
(and hence import) the same target twice.

 - Update the BlockDeviceZFS class to try use the -f option for imports
   that are part of a pacemaker_operation
 - Create plumbing to carry flags through the stack to the BlockDevice
   class
 - Update tests to cover new cases

Change-Id: I9bf4d090c498d0140a20fab333406eb82d9c763f
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22690
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>