HYD-6265 zfs find_device_and_children() doesn't resolve persistent naming links
The method does not seem to behave correctly with zpools created with
device name (e.g. sda) or by-path link (e.g. /dev/disk/by-path/pci-...).

Solution is to retrieve the full paths of devices making up the zpool,
compare these with the device basenames that make up the zpool and
derive the full path of the base device that can be normalized.

  - implement methods for comparing output of zpool list with zvol
    full paths and base device name, lists can then be used to identify
    partitions and base devices
  - remove the find_normalized_ends() method that is no longer used
  - add and update relevant test_zfs.py unit tests to cover extra
    functionality

Change-Id: I3a21ba1d012116a85b5ab57218977754694d9513
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/22784
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>