HYD-6719 Volumes can loose sync with servers
The Volumes should only contain the usable leafs of the device hierarchy. 3.1
contains some many improvements in the way in which the plugin resources are
handled, but we are still working our way through this legacy code.

The more we work on the code the better it will get, but we have to pull it
into shape slowly and whilst we will push the cases further into the corners
it may be some time before we have a solution with no unresolved corner cases.

This patch contains some significant improvements.

 - The removal of some linux nodes was node happening because the parameters
   to remove_missing_devices were reversed
 - The agent did a partprobe durring the ZFS scan which meant the zfs might
   see different partitions to the raw device scan. This is fixed by
   partprobing before any of the 'linux' scan occurs
 - The linux scan records the current-state at the end of the scan meaning
   some changes that occur durring the scan can be missed
 - Changed the linux plugin on the agent so that every 10 minutes it does a
   'safety send' of the last scan (it does not new scan) so that the manager
   will sync correctly as the tree changes. This is a workaround to not
   evaluating the whole tree each time, but attempts to do that lead to
   unexpected test failures

Change-Id: Ic1966575d366f580f79460c15249a6f073ee00ca
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23545
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User