HYD-6808 Zpools imported read-only when formatting target
Pools can be imported read-only when formatting a lustre
target. This is a bug as BlockDeviceZfs.import_() should
not allow the situation to occur.

Zfs properties are used to detect the readonly property
of the zpool identified by the path. Confusingly the zpool
has different properties as reported by the 'zfs' and 'zpool'
utilities. The property that affects the ability to create
a ZfsDataset on a zpool is the one reported by 'zpool'.

Previously the 'zfs' readonly property has been used to
detect if a pool is imported as readonly and has resulted
in this bug.

  - Create a BlockDeviceZfs.zpool_properties() method
  - Use said method when detecting if zpool is readonly
  - Fixup unit tests in test_manage_target and
    test_blockdevice_zfs
  - Remove duplicate testing of blockdevice class in
    test_manage_target

Change-Id: I8133497b9aa0970bd2e88840cce0f1b270811ebb
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/23709
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>