HYD-6810 Resolve inconsistencies in agent linux plugin device queries
RpcErrors resulting from KeyError exceptions received intermittently in efs
and ssi test runs.

Issue is occurring due to inconsistent data being returned by the agent during
*agent_session_start*. RpcErrors are appearing in *session_resource_add_parent*
and *path_to_major_minors* for what seems to be the same reason.
Zfsdatasets part of the data contain drives that don't exist within
the "devs" data and when the data is read and processed by the
resource manager, these errors occur.

Resolution involves zfs retrieving the major/minor identifiers once and reusing
the same data so as not to later use facts that the "devs" data did not discover.
In summary we want to scan for the data once per session on the agent
and use only one copy of the source data throughout the agent plugin,
this will avoid having inconsistencies between different parts of the data set.

  - Update calls which retrieve major_minor identifiers to point
    to public accessor methods which access BlockDevices.node_block_devices
    which was once populated in BlockDevices.__init__
  - Upkeep convention of using underscore prefix only for private methods
  - Move relevant parts of DeviceHelper to where they are used, (mainly
    BlockDevices) and remove empty file, move test_device_helper tests to
    test_linux
  - Move BlockDevices to device_plugins/linux_components/block_devices.py
  - When retrieving mm in path_to_major_minor, normalise device path
    before retrieving from node_block_devices dict
  - Move human_to_bytes utility method to chroma_common/lib/util.py
  - Keep all *PATH constants in BlockDevices and use as class variables
  - Fixup unit tests for different devices to address changes in BlockDevices

Change-Id: I2856c997279ed99b57c69280178e0812e4a32bac
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/23765
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>