HYD-6894 Filesystem Type for a target on a ZfsDataset is blank
Dataset filesystem_type should be set to 'zfs' when populated by the
linux device plugin in zfs.py.

Whilst a zfs dataset doesn't really contain a zfs filesystem - it is
a very good analogy which means that when a user views information
about the device they are told it is 'zfs' rather than 'ldiskfs'.
Whilst it is possible for a block device to have no filesystem and so
be nominally unusable, a dataset is always useable as a zfs device.

  - When reporting ZfsDatasets, set device 'filesystem_type' to 'zfs'
  - Minor syntax improvement for dict access

Change-Id: I4543934b60b897b9a1f1a9b89577744db824caef
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: https://review.whamcloud.com/26806
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>