HYD-6911 Test failure: zpool export - "no such pool"
There is a continous possibility of different threads accessing pools
an interfereing with each other.  The most likely failure is that one
pool exports a zpool whilst another is accessing it. ExportedZfsDevice
attempted to resolve this but was only called for cases where the
pool was being imported.

A more resilient approach is for anything that access a pool to take a
lock on the pool before doing so meaning that all zpool actions become
atomic.

This patch renames ExportedZfsDevice to be ZfsDevice and adds a parameter
so the caller can decide to import or not if the device is not imported.

Using this in a with contruction means that we have atomic zpool operations.
with has been added to all these operations.

A critical non-side effect of this is that the commands issued have not
changed (other than blkid which I was unable to resist simplifying)

 - Renamed ExportedZfsDevice to ZfsDevice
 - Add ability to lock already-imported device
 - Add tests to test new functionality
 - Enclose all zfs operations in a with ZfsDevice
 - ZfsDevice now provides itself to the with rather than a simple
   flag.

Change-Id: Ic6c3c17272b870899f12902c7e303c355f21b500
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: https://review.whamcloud.com/24222
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>