HYD-7081 Increase corosync token timeout to 17000ms
The Solution Architect group has provided us with a requirement
to use a default corosync totem timeout of 17000ms. This is in
response to some sites that have experienced unnecessary failover
events when their MDS is under enough load the totem doesn't get
scheduled in the CPU for more than the totem timeout.

Increasing the totem to 17000s will havethe consequent that real
failures take longer to detect and begin the failover process.
The field has indicated that our customer base will not mind the
delay.

In the future b4_0 release, IML will make this setting a tunable,
so users can change this up and down to suit their personal
preference of time to detect failure and risk of false failover,
based on their specific workloads.

  - Update the references to the totem timeout to 17000ms.

Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Change-Id: I11d3f86830735b2826a84f759ad56bbdd8496012
Reviewed-on: https://review.whamcloud.com/25071
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>