HYD-7043 don't allow configurations where stonith-enabled == false
We currently set stonith-enabled to false on Pacemaker setup.
Due to this, we open ourselves to corruption if power control has
not been setup prior to creating targets.

This patch raises alerts when stonith-enabled is false or None
and blocks target creation.

It also sets stonith-enabled to true by default

   - Raise an alert when stonith-enabled == false.
   - Block target creation when alert has been raised.
   - Add stonith-enabled prop to corosync scanning.

Change-Id: Ie627b156c8b97f2507fc626861151448a1bc5815
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: https://review.whamcloud.com/25454
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>