HYD-7213 Exceptions found in agent logs for known conditions (#74)
We should be able to rely on finding any exceptions in the chroma agent
logs as being indications of defects.  We have even periodically
discussed having the test framework fail tests if exceptions are found
in the agent log(s).

This unfortunately is not quite true (yet).

There are expected exceptions when the agent scan starts before
Pacemaker has been installed.  Rather than letting these pollute the
logs, we should catch them.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>