Integrate device-scanner in IML agent (#306)
* Integrate device-scanner into agent.

This patch integrates the device-scanner repo into the iml-agent. It is a replacement for the BlockDevices class and imperative lookup of devices.

Instead, a device map is held in a daemon, and updated whenever a udev event is received over a unix domain socket.

As a first pass, the BlockDevices class is updated to query state from the scanner so it is still polling (but quicker). The ultimate goal will be to move from a poll model to a push model which will necessitate further changes downstream.