Remove cluster-sim (#435)
* Remove cluster-sim

Now that we have a method to run tests locally via Vagrant / libvirt, we don't have as much need for the cluster-sim.

it is no longer used for GUI development either as this is done with a local Vagrant box.

Finally, we don't have as much bandwidth to actively maintain it, and upcoming changes will make it more difficult to work with.

Remove the cluster-sim.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Remove sim references.