Remove agent stop/disable/enable/start from RPM + 4.x -> 4.x Upgrades (#513)
* Remove agent stop/disable/start from RPM

Stopping the agent during upgrades kills the upgrade dead
in it's tracks.

The adminstrator, or the IML Manager are responsible for
restarting the agent after an agent upgrade.

The enabling of the unit in %post really ought to be replaced
using a preset and the %systemd_post RPM macro.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>

* Upgrade from both 3.x and 4.x

If the environment varialbe $UPGRADE_FROM_3x == true, then upgrade from
the last 3.x release otherwise (i.e. the default) upgrade from the most
recent 4.x release.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>

* Revert use of RHEL 7.5 since we are not using RHEL

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>

* Remove commented out code; fix backslash

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>