Create service grouping systemd units (#574)
Fixes #572

Now that we are creating smaller stand-alone services, we need the ability to start / stop them in tandem.

We can create either a iml-storage-server.target or dummy iml-storage-server.service unit to control starting / stopping these services in tandem.

This will handle the following services (so far) either directly or indirectly:

chroma-agent
device-scanner
scanner-proxy
mount-emitter
Things we need to consider:

We should repopulate the scanner if the service has stopped, but not if it has been running since boot.

Signed-off-by: Joe Grund <joe.grund@intel.com>