Remove caching of localMounts (#593)
* Remove caching of localMounts

localMounts are currently being cached at module load time, and are
never evaluated afterwards.

This is causing us a bunch of issues where mount state is not as
expected.

Remove the caching, so we evaulate local mounts each time we call the
function.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Update call

Update call in lustre.py as well.

* fixups

Signed-off-by: Joe Grund <joe.grund@intel.com>