Do not perform update to packages during IML install
IML has an action to install_packages. install_packages does an install, followed by an update as seen here:

https://github.com/whamcloud/iml-agent/blob/d1b688700a8d7878625baad967c830c942f9e806/chroma_agent/action_plugins/agent_updates.py#L113-L121

The update seems like an overreach for IML, even in managed mode as it doesn't appear to serve a purpose for IML managed repos.

The install_packages action is called on the manager side here:

https://github.com/whamcloud/integrated-manager-for-lustre/blob/054a091ed7eb7e3b001cc023feffd1adab493fae/chroma_core/models/host.py#L695-L698

and here:

https://github.com/whamcloud/integrated-manager-for-lustre/blob/054a091ed7eb7e3b001cc023feffd1adab493fae/chroma_core/models/host.py#L1415-L1419

In either case, a {yum,dnf} install should trigger an update if needed, so we should be ok to remove this.

Fixes #26

Signed-off-by: Joe Grund <jgrund@whamcloud.io>