Format Repo with black
Use [black](https://github.com/ambv/black)
to reformat the entire iml-agent repo.

Black is an autoformatter similar to gofmt, rustfmt, prettier et. all.

The caveat is black only runs on Python 3, but the upshot is it parses
Python 3 syntax, so after this change all the code in this repo is
parsable by Python 3.

Also add a travis step to check for consistent formatting in every
commit.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>