Create Mount Only When Necessary (#105)
* Create Mount Only When Necessary

The database does not currently reflect the actual state of mounts, so if it says the state is mounted, it could
actually be unmounted. The device scanner will eventually resolve this issue, but until then, we should check to see if
a mount exists before creating the mount. This will allow us to attempt to create a mount, even if the mount already
exists (in which case it would be a noop). This is useful when attempting to stream stratagem results to the client. We
can ensure that the client is mounted before attempting to stream.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* - fix broken test

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* - Use device scanner to get mount information

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* - Update test

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* - cr update

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* - remove dead code

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* - Cr change

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>