Handle single node resources when finding resource constraint
In the case of a resource configured on a single node, the result of
`cibxpath("query", '//constraints/rsc_location[@rsc="{}"]'.format(ha_label))` will be a single node
instead of a top-level node containing child nodes.

In this case, we should check the single node directly and fallback to
fetching children if the top-level node is not `rsc_location`.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>